+++ src/App.css
@@ -64,3 +64,7 @@ svg {
 .foodItem p {
   margin: 10px 0;
 }
+
+.favourites {
+  display: flex;
+}
